requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
attrs==18.2.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
Django==2.2.2
django-rest-knox==4.1.0
django-tinymce4-lite==1.7.5
djangorestframework==3.9.4
entrypoints==0.3
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.13.3
Jinja2==2.10
jsmin==2.2.2
jsonschema==3.0.0
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kiwisolver==1.0.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.1
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
pickleshare==0.7.5
pipenv==2018.11.26
prometheus-client==0.6.0
prompt-toolkit==2.0.9
pycparser==2.19
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
pyzmq==18.0.0
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.25.3
virtualenv==16.6.0
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.1
